dialects: (stream) remove unused stream.read and stream.write ops
