+++ tests/interpreters/test_riscv_snitch_interpreter.py
@@ -5,10 +5,9 @@ from xdsl.interpreter import Interpreter
 from xdsl.interpreters.func import FuncFunctions
 from xdsl.interpreters.riscv import RiscvFunctions
 from xdsl.interpreters.riscv_snitch import RiscvSnitchFunctions
+from xdsl.interpreters.stream import Acc, Nats
 from xdsl.ir import BlockArgument
 from xdsl.utils.test_value import TestSSAValue
-
-from .test_stream_interpreter import Acc, Nats
 
 
 def test_read_write():
